third_party/blink/web_tests/platform/generic/fast/dom/Window/es52-globals-expected.txt

PASS window.hasOwnProperty("Element") is true
PASS window.hasOwnProperty("x") is true
PASS window.hasOwnProperty("y") is false
FAIL window.hasOwnProperty("f") should be true. Was false.
FAIL window.hasOwnProperty("div") should be true. Was false.
FAIL window.hasOwnProperty("a") should be true. Was false.
PASS Element is not undefined
PASS x is 1
PASS y is undefined.
PASS f is undefined.
PASS div is undefined.
PASS a is undefined.
PASS successfullyParsed is true

TEST COMPLETE
